更新

+++ app/src/main/java/com/exp/memoria/data/local/entity/RawMemory.kt
@@ -20,7 +20,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "raw_memory")
 data class RawMemory(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0, // [cite: 55]
+    val id: Long = 0, // [cite: 55]
     val user_query: String, // [cite: 56]
     val llm_response: String, // [cite: 57]
     val timestamp: Long // [cite: 58]
